refactored more and added views; straightened out and now at a stable version.
